♻️refactor : decode메서드 Decodable 확장해 통합
